Add SimpleJdbcEventListener with more coarse grained callbacks
